records: merge authors/receivers into receivers
