+++ inspirehep/modules/authors/__init__.py
@@ -25,4 +25,3 @@
 from __future__ import absolute_import, division, print_function
 
 from .ext import INSPIREAuthors  # noqa: F401
-from .receivers import *         # noqa: F401,F403
